Cleaned up controllers, improved logics on wallet operations, added tests

+++ app/Http/Controllers/WalletController.php
@@ -5,75 +5,57 @@ namespace App\Http\Controllers;
 use App\Http\Requests\WalletStoreRequest;
 use App\Http\Requests\WalletUpdateRequest;
 use App\Models\Wallet;
+use Illuminate\Contracts\Foundation\Application;
+use Illuminate\Contracts\View\Factory;
+use Illuminate\Contracts\View\View;
+use Illuminate\Http\RedirectResponse;
+use Illuminate\Routing\Redirector;
 use Illuminate\Support\Facades\Auth;
 
 class WalletController extends Controller
 {
-
-    /**
-     * WalletController constructor.
-     */
-    public function __construct()
+    public function index(): Factory|View|Application
     {
-        $this->middleware('auth');
-    }
-
-    /**
-     * Display a listing of the resource.
-     *
-     * @return \Illuminate\Contracts\View\Factory|\Illuminate\View\View
-     */
-    public function index()
-    {
-        $wallets = Auth::user()->wallets()->latest()->get();
+        $wallets = Auth::user()
+            ->wallets()
+            ->latest()
+            ->get();
 
         return view('wallets.index', compact('wallets'));
     }
 
-    /**
-     * Show the form for creating a new resource.
-     *
-     * @return \Illuminate\Contracts\View\Factory|\Illuminate\View\View
-     */
-    public function create()
+    public function create(): Factory|View|Application
     {
         return view('wallets.create');
     }
 
-    /**
-     * Store a newly created resource in storage.
-     *
-     * @param WalletStoreRequest $request
-     * @return \Illuminate\Http\RedirectResponse|\Illuminate\Routing\Redirector|void
-     */
-    public function store(WalletStoreRequest $request)
+    public function store(WalletStoreRequest $request): Redirector|Application|RedirectResponse
     {
-        $wallet = $request->validated();
-        $wallet['user_id'] = Auth::user()->id;
-        Wallet::create($wallet);
+        Auth::user()
+            ->wallets()
+            ->create($request->validated());
 
         return redirect(route('wallets.index'))
             ->with('alert', ['type' => 'success', 'message' => 'Wallet created successfully']);
     }
 
     /**
-     * Show the form for editing the specified resource.
-     *
-     * @param \App\Models\Wallet $wallet
-     * @return \Illuminate\Contracts\View\Factory|\Illuminate\Http\Response|\Illuminate\View\View
+     * @throws \Illuminate\Auth\Access\AuthorizationException
      */
-    public function edit(Wallet $wallet)
+    public function edit(Wallet $wallet): Factory|View|Application
     {
+        $this->authorize('edit', $wallet);
+
         return view('wallets.edit', compact('wallet'));
     }
 
     /**
-     * @param WalletUpdateRequest $request
-     * @param Wallet              $wallet
-     * @return \Illuminate\Contracts\Foundation\Application|\Illuminate\Http\RedirectResponse|\Illuminate\Routing\Redirector
+     * @throws \Illuminate\Auth\Access\AuthorizationException
      */
-    public function update(WalletUpdateRequest $request, Wallet $wallet)
+    public function update(WalletUpdateRequest $request, Wallet $wallet): Redirector|Application|RedirectResponse
     {
+        $this->authorize('update', $wallet);
+
         $wallet->fill($request->validated())->save();
 
         return redirect(route('wallets.edit', $wallet))
@@ -81,14 +63,12 @@ class WalletController extends Controller
     }
 
     /**
-     * Remove the specified resource from storage.
-     *
-     * @param \App\Models\Wallet $wallet
-     * @return \Illuminate\Http\RedirectResponse|\Illuminate\Http\Response|\Illuminate\Routing\Redirector
-     * @throws \Exception
+     * @throws \Illuminate\Auth\Access\AuthorizationException
      */
-    public function destroy(Wallet $wallet)
+    public function destroy(Wallet $wallet): Redirector|Application|RedirectResponse
     {
+        $this->authorize('delete', $wallet);
+
         $wallet->delete();
 
         return redirect(route('wallets.index'))
